Make updateAgedBrie an extension method
